requirements.txt
accelerate==1.2.1
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asgiref==3.8.1
blobfile==3.0.0
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
diskcache==5.6.3
distro==1.9.0
Django==5.1.4
django-environ==0.12.0
filelock==3.16.1
fire==0.7.0
fsspec==2024.12.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
llama_cpp_python==0.3.5
llama_models==0.0.63
llama_stack==0.0.63
llama_stack_client==0.0.63
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
openai==1.58.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==6.1.1
pyaml==25.1.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.0
PyMuPDF==1.25.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3transfer==0.11.1
safetensors==0.5.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
youtube-dl==2021.12.17
